refactor(app): move citation/reference subsystem into citation/
Moves cite.ts, bibtex2csl.ts, doi2csl.ts, reference.tsx,
reference-list.tsx, how-to-cite.tsx, terms-of-use.tsx and
terms-of-use-developer.tsx out of the flat shared/ folder into a new
citation/ folder, since they form a cohesive citation/reference
subsystem used together by the author set edit pages, scat-cs/scat-css
inspect pages and the developer page. Updates all import paths
accordingly. No behavioral changes.
