+++ app/src/app/api/scat-cs/inspect/legacy/route.ts
@@ -5,7 +5,7 @@
 import { annotateMixture } from "@/shared/annotate-mixture";
 import { db } from "@lxcat/database";
 import { okResponse } from "../../../../../shared/api-responses";
-import { reference2bibliography } from "../../../../../shared/cite";
+import { reference2bibliography } from "@/citation/cite";
 import { mapObject } from "../../../../../shared/utils";
 import {
   hasDeveloperOrDownloadRole,
