+++ app/src/app/scat-cs/compute/page.tsx
@@ -8,7 +8,7 @@ import { db } from "@lxcat/database";
 import { LTPMixtureWithReference } from "@lxcat/schema";
 import Script from "next/script";
 import { z } from "zod";
-import { reference2bibliography } from "../../../shared/cite";
+import { reference2bibliography } from "@/citation/cite";
 import { mapObject } from "../../../shared/utils";
 import { IdsSchema } from "../ids-schema";
 import { BolsigPage, BolsigPageProps } from "./bolsig-page";
