+++ app/src/app/scat-cs/inspect/legacy-action.ts
@@ -5,7 +5,7 @@
 "use server";
 
 import { annotateMixture } from "@/shared/annotate-mixture";
-import { reference2bibliography } from "@/shared/cite";
+import { reference2bibliography } from "@/citation/cite";
 import { mapObject } from "@/shared/utils";
 import { convertMixture } from "@lxcat/converter";
 import { LTPMixture } from "@lxcat/schema";
